refactor: extract poll creation form logic to custom hook
